+++ versions.props
@@ -2,7 +2,7 @@ com.fasterxml.jackson.*:* = 2.19.1
 com.google.auto.service:* = 1.1.1
 com.google.errorprone:error_prone_* = 2.49.0
 com.google.guava:guava = 33.6.0-jre
-com.netflix.nebula:nebula-test = 10.6.2
+com.palantir.baseline:gradle-baseline-java = 7.4.0
 com.palantir.gradle.auto-parallelizable:* = 1.4.0
 com.palantir.safe-logging:* = 3.14.0
 com.palantir.suppressible-error-prone:* = 2.28.0
